+++ telco_si/migrations/20260919200715_initial_schema.sql
@@ -1,0 +1,2 @@
+-- Initial database migration.
+-- Business tables will be introduced by subsequent migrations.